requirements.txt
functions-framework==3.*
flask>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
pytz>=2021.1
google-cloud-storage>=2.0.0
google-cloud-secretmanager>=2.0.0
gunicorn>=20.1.0

